Added another test
