+++ src/shared/options/augmentOptionsWithExcludes.test.ts
@@ -3,10 +3,14 @@ import { describe, expect, it, vi } from "vitest";
 import { Options } from "../types.js";
 import { augmentOptionsWithExcludes } from "./augmentOptionsWithExcludes.js";
 
+const mockMultiselect = vi.fn();
 const mockSelect = vi.fn();
 
 vi.mock("@clack/prompts", () => ({
 	isCancel: () => false,
+	get multiselect() {
+		return mockMultiselect;
+	},
 	get select() {
 		return mockSelect;
 	},
@@ -56,16 +60,49 @@ const optionsBase = {
 } satisfies Options;
 
 describe("augmentOptionsWithExcludes", () => {
-	it("prompts for base when no exclusions are provided", async () => {
+	it("returns options directly when no exclusions are provided and 'base' is provided for the prompt", async () => {
 		const base = "everything";
 
-		mockSelect.mockResolvedValue(base);
+		mockSelect.mockResolvedValueOnce(base);
 
 		const actual = await augmentOptionsWithExcludes(optionsBase);
 
 		expect(actual).toEqual({
 			...optionsBase,
 			base,
+		});
+	});
+
+	it.only("returns options based on the select when no exclusions are provided and 'prompt' is provided for the prompt", async () => {
+		const base = "prompt";
+
+		mockSelect.mockResolvedValueOnce(base);
+		mockMultiselect.mockResolvedValue([]);
+
+		const actual = await augmentOptionsWithExcludes(optionsBase);
+
+		expect(actual).toEqual({
+			...optionsBase,
+			base,
+			excludeAllContributors: true,
+			excludeCompliance: true,
+			excludeLintDeprecation: true,
+			excludeLintESLint: true,
+			excludeLintJSDoc: true,
+			excludeLintJson: true,
+			excludeLintKnip: true,
+			excludeLintMd: true,
+			excludeLintPackageJson: true,
+			excludeLintPackages: true,
+			excludeLintPerfectionist: true,
+			excludeLintRegex: true,
+			excludeLintSpelling: true,
+			excludeLintStrict: true,
+			excludeLintStylistic: true,
+			excludeLintYml: true,
+			excludeReleases: true,
+			excludeRenovate: true,
+			excludeTests: true,
 		});
 	});
